Move libp2p sync testing a separate directory
